packages.json is not needed for detection
Remove a line that stated that packages.json is required for detection of the
correct buildpack.

This example specifies the buildpack in the manifest.yml, which means that
buildpack auto-detection is bypassed.

We a decision to do this in our examples, to decrease the time taken to stage
applications, as early feedback from people who used our CF was that this felt
slow. There's more detail about buildpack detection and its effect on
performance here: https://docs.cloudfoundry.org/buildpacks/detection.html

We could review the effect of buildpack detection on performance at a future
time, but for now this line makes no sense.