Remove Flask Documentation
Flask documentation is currently not required by our users. We
are removing it as documentation is currently being worked on.

Once we have more documentation in place and users express a need for
flask documentation we will put new guidance in place. This will
allow us to produce more consistent useful documentation.