Add warning about API downtime during updates in beta
Story: Document Downtime for APIs during Upgrades

https://www.pivotaltracker.com/story/show/127792551

"We need to let tenants know the effects of our releases on the API,
and the sort of errors they may experience.
This is important from a user experience point of view and from a user
Continuous Deployment perspective."

On discussion, not clear if we can predict all errors, so just added a
general warning that valid commands may fail for a brief 5 minute
period.