Document that custom builldpacks are not supported
Document Custom Buildpack decision

https://www.pivotaltracker.com/story/show/126403487

"Outline the fact that we don't currently support custom buildpacks and
the reason for this."

Decided just to say that they are not supported. Main reason for people
wanting them to use them is to get a newer version of the
runtime/framework, so mentioned that the normal build packs are updated
regularly.