Document that cf-ssh is deliberately disabled
https://www.pivotaltracker.com/story/show/126051461
"The option to SSH into an app is disabled until we determine if there
is a need for it. There are possible reasons why it is better that
tenants don't use it. The error message you get when you try it is not
helpful, and tenants may read about it in CF docs, so out docs should
explain why it does not work."