More complete list of commands that cause 404s
@dcarley pointed out in https://github.com/cloudfoundry/cli/issues/922
that cf restage also causes downtime.

I included that and `cf scale`, which is OK to run when changing number of app
instances, but causes downtime when changing vertical scaling limits (memory
and disk) or forcing a restart.

Also reworded to say that anything that restarts the app should be considered
unsafe.