Remove duplicate limitations page
This was left behind by accident in #17 and only contains information that is
in docs/getting_started/limitations.md

In 7dab157 we moved the file to the getting started section of the docs.  Then
in b70cf5e we reverted that. The text about Docker being disabled was added to
the new limitations page in 4c21b00.

We discussed briefly in slack if we needed to redirect the old page to the new
and decided it was unlikely that many people would have links saved that
pointed to the old page.