Document IP whitelist restrictions
This update adds an explanation of the IP whitelist to limit CLI access
to PaaS. See Pivotal story below for details.

If tenant's IP is not whitelisted, they can't use the PaaS at all, and
this is known to have been a problem for several users, so I have added
warnings in the Quick Setup guide, the Troubleshooting section, and in
the Limitations section.

To test: check that explanation is accurate and easy to understand.
Check that link to whitelist is accurate and confirm we should point
users to this information. In "Troubleshooting", check that the
description of what happens if you try to use the client without being
whitelisted is correct.

Pivotal story:
#131354965

Currently we only allow a limited set of IP addresses to access the CF
endpoints (CF, loggregator, UAA).
This is a temporary way of reducing our attack surface until we are
able to do some more work to secure the system from attack.
Things we could and possibly should document for tenants:
that their deployed applications are public and not whitelisted
that the restrictions will prevent them from running cf commands from
an IP address not on the list
that this is a temporary measure
that we can add IP addresses to the whitelist if required by tenants