Documentation on creating/managing spaces
Provides better guidance on creating spaces and assigning developer
role to use a space.

To test: check that provided instructions work and the information
provided is accurate and everything that's needed for basic space
management.

Pivotal #133535113

As an user I need to understand how to create spaces on the PaaS and
who has permissions to do so, so that my team can create and manage
spaces effectively.
Acceptance Criteria
Documentation includes information on how to create spaces (and who has
permissions to do so). This should also include more detailed info on
setting up a space (including what the default settings are and how
permissions work)
User Research
During October lab sessions users had difficulty finding the correct
command to create a new space on the PaaS. We currently link off to the
CF documentation on this - users got lost and didn’t find what they
were looking for.