Add "Names, Routes and Domains" section
Add a section explaining how cloudapps.digital hostnames work, how to
avoid name clash. Document that currently the only way to use your own
custom domain is with a CDN (mentioned Limitations and in the new
section).

To test: check that commands are accurate. Check that these
recommendations are correct.

Pivotal story: #133534409

As a user I need to understand how to prevent routing clashes so that I
can deploy the same application in multiple spaces on the PaaS.
Acceptance Criteria
Users understand how to prevent routing errors, including best practice
naming conventions and how to use hostnames or domains (whichever is
the team’s preferred method).
User Research
During lab testing users encountered an error when they tried to deploy
the same application to different spaces in an org. This happens across
orgs as well.