Warn users not to set PORT; troubleshooting if they have
Changes to address DeskPRO ticket 1798 where a tenant had problems
caused by the PORT variable being changed from the system-provided
value. We are not clear how that happened.

Fixes:
* in env_variables, put more information about system provided
variables and say not to change them yourself; provide link to CF docs
about this early in relevant section instead of under a separate
"Further reading" heading
* add Troubleshooting page about how to fix it if PORT gets set

To test:

check content, especially that steps to unset PORT are correct and can
be run by a tenant rather than the PaaS team