Work around broken search integration
MkDocs and Readthedocs don't currently agree on how to do search.
This means that search has poor UX on the publically deployed docs
on readthedocs.io.

Here's a workaround suggested in github comments that we could use
to fix things for us while we wait for the two projects to reconcile.

See:
https://github.com/rtfd/readthedocs.org/pull/2289
https://github.com/rtfd/readthedocs.org/issues/1088