Redirect readthedocs to docs.cloud.service.gov.uk
Readthedocs doesn't allow redirections, or custom themes that would
allows us to deploy a simpler solution, so this replaces every page in
the documentation with a block of text linking to the correct page in
the new documentation.

Automatic redirection is achieved because RTD allows you to include
custom javascript, so this simply grabs the link in the page and changes
the window location after 5 seconds.